feat: PDF generation, device returns, delegation, budget tracking, SLA/contract fixes
- Add PDF generation for allocation letters, contract summaries, order confirmations (pdfkit, DOJCD letterhead)
- Add device return lifecycle (Requested → Completed) with stock restoration on completion
- Add approval delegation with active-delegation middleware (requireManagerOrDelegate)
- Add department budget tracking with monthly ceiling and live spend comparison
- Consolidate all 6 migrations + department-isolation script into all_migrations.sql
- Fix SLA report 500: add ::integer casts to CASE params in fetchSlaData()
- Fix contract summary aliases (total_contracts→total, expiring_30_days→expiring_30, etc.)
- Fix getEligibleDelegates: use is_deleted=FALSE instead of non-existent is_active
- Wire all 4 new route groups into app.js; extend approverRoutes with delegation middleware

+++ src/services/reportService.js
@@ -136,9 +136,9 @@ async function fetchSlaData() {
                         (SELECT ap.approval_date FROM approval ap WHERE ap.application_id = a.application_id AND ap.approval_stage = 'finance' LIMIT 1)
                 END AS stage_entry_date,
                 CASE
-                    WHEN a.application_status = 'Pending'         THEN $1
-                    WHEN a.application_status = 'Pending_Finance' THEN $2
-                    WHEN a.application_status = 'Approved'        THEN $3
+                    WHEN a.application_status = 'Pending'         THEN $1::integer
+                    WHEN a.application_status = 'Pending_Finance' THEN $2::integer
+                    WHEN a.application_status = 'Approved'        THEN $3::integer
                 END AS sla_days
             FROM application a
             JOIN client_user cu ON a.client_user_id = cu.client_user_id
